+++ pkg/dashboard/subproc/app.go
@@ -1,15 +1,11 @@
 package subproc
 
 import (
-	"fmt"
 	"github.com/joomcode/errorx"
-	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	"helm.sh/helm/v3/pkg/action"
 	"helm.sh/helm/v3/pkg/cli"
 	"helm.sh/helm/v3/pkg/registry"
-	"helm.sh/helm/v3/pkg/release"
-	v1 "k8s.io/apimachinery/pkg/apis/testapigroup/v1"
 	"os"
 
 	// Import to initialize client auth plugins.
@@ -26,12 +22,10 @@ type HelmNSConfigGetter = func(ns string) (*action.Configuration, error)
 type Application struct {
 	HelmConfig HelmNSConfigGetter
 
-	K8s                *K8s
-	CurrentContextName string
-	Scanners           []Scanner
+	K8s *K8s
 
-	releases     []*Release
-	repositories []*Repository
+	Releases     *Releases
+	Repositories *Repositories
 }
 
 func NewApplication(helmConfig HelmNSConfigGetter) (*Application, error) {
@@ -48,28 +42,13 @@ func NewApplication(helmConfig HelmNSConfigGetter) (*Application, error) {
 	return &Application{
 		HelmConfig: helmConfig,
 		K8s:        k8s,
+		Releases: &Releases{
+			HelmConfig: helmConfig,
+		},
+		Repositories: &Repositories{
+			HelmConfig: helmConfig,
+		},
 	}, nil
-}
-
-func (a *Application) GetReleases() ([]*Release, error) {
-	hc, err := a.HelmConfig("")
-	if err != nil {
-		return nil, errorx.Decorate(err, "failed to get helm config for namespace '%s'", "")
-	}
-
-	client := action.NewList(hc)
-	client.All = true
-	client.AllNamespaces = true
-	client.Limit = 0
-	releases, err := client.Run()
-	if err != nil {
-		return nil, errorx.Decorate(err, "failed to get list of releases")
-	}
-	a.releases = []*Release{}
-	for _, r := range releases {
-		a.releases = append(a.releases, &Release{HelmConfig: a.HelmConfig, Orig: r})
-	}
-	return a.releases, nil
 }
 
 func (a *Application) CheckConnectivity() error {
@@ -83,21 +62,6 @@ func (a *Application) CheckConnectivity() error {
 		return errorx.Decorate(err, "failed to validate k8s cluster connectivity")
 	}
 	return nil
-}
-
-func (a *Application) ReleaseByName(namespace string, name string) (*Release, error) {
-	rels, err := a.GetReleases()
-	if err != nil {
-		return nil, errorx.Decorate(err, "failed to get list of releases")
-	}
-
-	for _, r := range rels {
-		if r.Orig.Namespace == namespace && r.Orig.Name == name {
-			return r, nil
-		}
-	}
-
-	return nil, errors.New(fmt.Sprintf("release '%s' is not found in namespace '%s'", name, namespace))
 }
 
 func NewHelmConfig(ctx string, ns string) (*action.Configuration, error) {
@@ -127,74 +91,4 @@ func NewHelmConfig(ctx string, ns string) (*action.Configuration, error) {
 	}
 
 	return actionConfig, nil
-}
-
-type Release struct {
-	HelmConfig HelmNSConfigGetter
-	Orig       *release.Release
-	revisions  []*Release
-}
-
-func (r *Release) History() ([]*Release, error) {
-	hc, err := r.HelmConfig(r.Orig.Namespace)
-	if err != nil {
-		return nil, errorx.Decorate(err, "failed to get helm config for namespace '%s'", "")
-	}
-
-	client := action.NewHistory(hc)
-	revs, err := client.Run(r.Orig.Name)
-	if err != nil {
-		return nil, errorx.Decorate(err, "failed to get revisions of release")
-	}
-
-	r.revisions = []*Release{}
-	for _, rev := range revs {
-		r.revisions = append(r.revisions, &Release{HelmConfig: r.HelmConfig, Orig: rev})
-	}
-
-	return r.revisions, nil
-}
-
-func (r *Release) Uninstall() error {
-	hc, err := r.HelmConfig(r.Orig.Namespace)
-	if err != nil {
-		return errorx.Decorate(err, "failed to get helm config for namespace '%s'", "")
-	}
-
-	client := action.NewUninstall(hc)
-	_, err = client.Run(r.Orig.Name)
-	if err != nil {
-		return errorx.Decorate(err, "failed to uninstall release")
-	}
-	return nil
-}
-
-func (r *Release) Rollback(toRevision int) error {
-	hc, err := r.HelmConfig(r.Orig.Namespace)
-	if err != nil {
-		return errorx.Decorate(err, "failed to get helm config for namespace '%s'", "")
-	}
-
-	client := action.NewRollback(hc)
-	client.Version = toRevision
-	return client.Run(r.Orig.Name)
-}
-
-func (r *Release) ParsedManifests() ([]*v1.Carp, error) {
-	carps, err := ParseManifests(r.Orig.Manifest)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, carp := range carps {
-		if carp.Namespace == "" {
-			carp.Namespace = r.Orig.Namespace
-		}
-	}
-
-	return carps, err
-}
-
-type Repository struct {
-	// TODO
 }
